+++ packages/react/src/Screen/Screen.tsx
@@ -10,12 +10,22 @@ import type { ForwardedRef, HTMLAttributes, PropsWithChildren } from 'react'
 type ScreenMaxWidth = 'wide' | 'x-wide'
 
 export type ScreenProps = {
+  /** Whether the screen should stretch to the full height of the viewport. */
+  fullHeight?: boolean
+  /** The maximum width of the screen. */
   maxWidth?: ScreenMaxWidth
 } & PropsWithChildren<HTMLAttributes<HTMLDivElement>>
 
 export const Screen = forwardRef(
-  ({ children, className, maxWidth = 'wide', ...restProps }: ScreenProps, ref: ForwardedRef<HTMLDivElement>) => (
-    <div {...restProps} ref={ref} className={clsx('ams-screen', `ams-screen--${maxWidth}`, className)}>
+  (
+    { children, className, fullHeight, maxWidth = 'wide', ...restProps }: ScreenProps,
+    ref: ForwardedRef<HTMLDivElement>,
+  ) => (
+    <div
+      {...restProps}
+      ref={ref}
+      className={clsx('ams-screen', fullHeight && 'ams-screen--full-height', `ams-screen--${maxWidth}`, className)}
+    >
       {children}
     </div>
   ),