+++ storybook/config/main.ts
@@ -30,6 +30,9 @@ const config: StorybookConfig = {
   docs: {
     autodocs: true,
   },
+  typescript: {
+    reactDocgen: 'react-docgen-typescript',
+  },
 }
 
 export default config