+++ storybook/src/components/Blockquote/Blockquote.stories.tsx
@@ -3,8 +3,9 @@
  * Copyright Gemeente Amsterdam
  */
 
-import { Blockquote } from '@amsterdam/design-system-react'
+import { Blockquote } from '@amsterdam/design-system-react/src'
 import { Meta, StoryObj } from '@storybook/react'
+import { inverseColorDecorator } from '../shared/decorators'
 import { exampleQuote } from '../shared/exampleContent'
 
 const quote = exampleQuote()
@@ -20,33 +21,18 @@ const meta = {
     children: {
       table: { disable: false },
     },
-    inverseColor: { control: 'boolean' },
   },
+  decorators: [inverseColorDecorator],
 } satisfies Meta<typeof Blockquote>
 
 export default meta
 
 type Story = StoryObj<typeof meta>
 
-export const Default: Story = {
-  decorators: [
-    (Story, context) => (
-      <div className={context.args.inverseColor ? 'ams-docs-dark-background' : undefined}>
-        <Story />
-      </div>
-    ),
-  ],
-}
+export const Default: Story = {}
 
 export const InvertedColor: Story = {
   args: {
     inverseColor: true,
   },
-  decorators: [
-    (Story, context) => (
-      <div className={context.args.inverseColor ? 'ams-docs-dark-background' : undefined}>
-        <Story />
-      </div>
-    ),
-  ],
 }