+++ storybook/src/components/OrderedList/OrderedList.stories.tsx
@@ -3,8 +3,9 @@
  * Copyright Gemeente Amsterdam
  */
 
-import { Heading, OrderedList, Paragraph } from '@amsterdam/design-system-react'
+import { Heading, OrderedList, Paragraph } from '@amsterdam/design-system-react/src'
 import type { Meta, StoryObj } from '@storybook/react'
+import { inverseColorDecorator } from '../shared/decorators'
 import { exampleOrderedList } from '../shared/exampleContent'
 
 const orderedListItems = exampleOrderedList().map((text, index) => (
@@ -22,18 +23,10 @@ const meta = {
     start: undefined,
   },
   argTypes: {
-    inverseColor: { control: 'boolean' },
-    markers: { control: 'boolean' },
     reversed: { control: 'boolean' },
     start: { control: 'number' },
   },
-  decorators: [
-    (Story, context) => (
-      <div className={context.args.inverseColor ? 'ams-docs-dark-background' : undefined}>
-        <Story />
-      </div>
-    ),
-  ],
+  decorators: [inverseColorDecorator],
 } satisfies Meta<typeof OrderedList>
 
 export default meta