+++ storybook/src/components/PageMenu/PageMenu.stories.tsx
@@ -3,13 +3,17 @@
  * Copyright Gemeente Amsterdam
  */
 
-import { PageMenu } from '@amsterdam/design-system-react'
-import { LoginIcon } from '@amsterdam/design-system-react-icons'
+import { PageMenu } from '@amsterdam/design-system-react/src'
+import { LoginIcon, SearchIcon } from '@amsterdam/design-system-react-icons'
 import { Meta, StoryObj } from '@storybook/react'
 
 const meta = {
   title: 'Components/Navigation/Page Menu',
   component: PageMenu,
+  args: {
+    alignEnd: false,
+    wrap: undefined,
+  },
 } satisfies Meta<typeof PageMenu>
 
 export default meta
@@ -29,9 +33,25 @@ export const Default: Story = {
   },
 }
 
-export const AlignEnd: Story = {
+export const Alignment: Story = {
   args: {
     alignEnd: true,
+    children: [
+      <PageMenu.Link href="#" key={1}>
+        Contact
+      </PageMenu.Link>,
+      <PageMenu.Link href="#" key={2}>
+        Mijn Amsterdam
+      </PageMenu.Link>,
+      <PageMenu.Link href="#" icon={SearchIcon} key={3}>
+        Zoeken
+      </PageMenu.Link>,
+    ],
+  },
+}
+
+export const Wrapping: Story = {
+  args: {
     children: [
       <PageMenu.Link href="#" key={1}>
         Onderzoeken