+++ storybook/src/components/TextArea/TextArea.stories.tsx
@@ -3,7 +3,7 @@
  * Copyright Gemeente Amsterdam
  */
 
-import { TextArea, TextAreaProps } from '@amsterdam/design-system-react'
+import { TextArea, TextAreaProps } from '@amsterdam/design-system-react/src'
 import { useArgs } from '@storybook/preview-api'
 import { Meta, StoryObj } from '@storybook/react'
 import { exampleParagraph } from '../shared/exampleContent'
@@ -35,11 +35,6 @@ const meta = {
     rows: {
       control: {
         type: 'number',
-      },
-    },
-    disabled: {
-      control: {
-        type: 'boolean',
       },
     },
   },